refactor: adjusted routing, breaks missing ones
