+++ routes/toys.tsx
@@ -1,12 +1,12 @@
 //? Create blog content inside Base component
-import { Base } from "../../components/Base.tsx";
+import { Base } from "../components/Base.tsx";
 //? Import a Card component that automatically creates
 //? redirection link, card title and centers card child
-import { CarouselCard } from "../../components/CarouselCard.tsx";
+import { CarouselCard } from "../components/CarouselCard.tsx";
 //? Import CustomHead with appropriate metadata
-import { CustomHead } from "../../components/CustomHead.tsx";
+import { CustomHead } from "../components/CustomHead.tsx";
 //? Navigation Buttons to go back to the previous page or to the next page (optional)
-import BlogNavigationButtons from "../../islands/BlogNavigationButtons.tsx";
+import BlogNavigationButtons from "../islands/BlogNavigationButtons.tsx";
 
 export default function Home() {
   return (
